# backend/openapi/api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Platform Backend",
    "description": "Health Platform & Digital Medicine Marketplace backend (modular monolith). **Identity**: registration, authentication, sessions, MFA, social login, and role/permission management. **Onboarding**: applicant onboarding, document review, verification checks, and maker-checker approval for doctors, nutritionists, hospitals, clinics, and pharmacies.",
    "version": "0.2.0"
  },
  "paths": {
    "/api/v1/auth/register/patient": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new patient",
        "operationId": "register_patient_api_v1_auth_register_patient_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_RegisterResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Email or mobile already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register/doctor": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register a new doctor applicant",
        "operationId": "register_doctor_api_v1_auth_register_doctor_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoctorRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_RegisterResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Email or mobile already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in with email and password",
        "description": "Returns tokens directly, or an MFA challenge if the account has MFA enabled.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Either a token pair or an MFA challenge",
            "content": {
              "application/json": {
                "schema": {},
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "accessToken": "...",
                        "refreshToken": "...",
                        "tokenType": "Bearer",
                        "expiresIn": 900,
                        "mfaRequired": false
                      }
                    }
                  },
                  "mfa_required": {
                    "value": {
                      "success": true,
                      "data": {
                        "mfaRequired": true,
                        "challengeId": "8476985a-6856-4057-8f9d-b6f5cde09427",
                        "methods": [
                          "AUTHENTICATOR"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "423": {
            "description": "Account locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Account suspended or deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Rotate a refresh token for a new access/refresh token pair",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope_TokenPair_"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalid, expired, revoked, or reused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out the current session",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/logout-all": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log out of all devices/sessions",
        "operationId": "logout_all_api_v1_auth_logout_all_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceInfoIn": {
        "properties": {
          "deviceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviceid"
          },
          "deviceName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devicename"
          },
          "platform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform",
            "description": "WEB | IOS | ANDROID | OTHER"
          },
          "appVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appversion"
          },
          "rememberMe": {
            "type": "boolean",
            "title": "Rememberme",
            "default": false
          }
        },
        "type": "object",
        "title": "DeviceInfoIn"
      },
...
      "DoctorRegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "mobileNumber": {
            "type": "string",
            "title": "Mobilenumber"
          },
          "firstName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Firstname"
          },
          "lastName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Lastname"
          },
          "termsAccepted": {
            "type": "boolean",
            "title": "Termsaccepted"
          },
          "privacyPolicyAccepted": {
            "type": "boolean",
            "title": "Privacypolicyaccepted"
          },
          "professionalDataConsentAccepted": {
            "type": "boolean",
            "title": "Professionaldataconsentaccepted"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "mobileNumber",
          "firstName",
          "lastName",
          "termsAccepted",
          "privacyPolicyAccepted",
          "professionalDataConsentAccepted"
        ],
        "title": "DoctorRegisterRequest"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInfoIn"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "PatientRegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "mobileNumber": {
            "type": "string",
            "title": "Mobilenumber"
          },
          "firstName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Firstname"
          },
          "lastName": {
            "type": "string",
            "maxLength": 150,
            "minLength": 1,
            "title": "Lastname"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "title": "Dateofbirth"
          },
          "termsAccepted": {
            "type": "boolean",
            "title": "Termsaccepted"
          },
          "privacyPolicyAccepted": {
            "type": "boolean",
            "title": "Privacypolicyaccepted"
          },
          "healthDataConsentAccepted": {
            "type": "boolean",
            "title": "Healthdataconsentaccepted"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "mobileNumber",
          "firstName",
          "lastName",
          "dateOfBirth",
          "termsAccepted",
          "privacyPolicyAccepted",
          "healthDataConsentAccepted"
        ],
        "title": "PatientRegisterRequest"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInfoIn"
          }
        },
        "type": "object",
        "required": [
          "refreshToken"
        ],
        "title": "RefreshTokenRequest"
      },
      "RegisterResponse": {
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "accountStatus": {
            "type": "string",
            "title": "Accountstatus"
          },
          "emailVerificationRequired": {
            "type": "boolean",
            "title": "Emailverificationrequired"
          },
          "mobileVerificationRequired": {
            "type": "boolean",
            "title": "Mobileverificationrequired"
          }
        },
        "type": "object",
        "required": [
          "userId",
          "accountStatus",
          "emailVerificationRequired",
          "mobileVerificationRequired"
        ],
        "title": "RegisterResponse"
      },
...
      "SuccessEnvelope_RegisterResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/RegisterResponse"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[RegisterResponse]"
      },
      "SuccessEnvelope_TokenPair_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/TokenPair"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessEnvelope[TokenPair]"
      },
...
      "TokenPair": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "refreshToken": {
            "type": "string",
            "title": "Refreshtoken"
          },
          "tokenType": {
            "type": "string",
            "title": "Tokentype",
            "default": "Bearer"
          },
          "expiresIn": {
            "type": "integer",
            "title": "Expiresin"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ],
        "title": "TokenPair"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token issued by /auth/login or /auth/refresh",
        "scheme": "bearer"
      }
    }
  },
...
}